refactor(handler): move schedule catalog surface to schedules_catalog.go (#1022)
Pure code motion: --color-moved clean, no exported-API or behaviour change.
